Move consolidate_attrs fn
