+++ shiny/experimental/ui/_valuebox.py
@@ -4,8 +4,8 @@ from typing import Callable, Optional
 
 from htmltools import Tag, TagAttrs, TagAttrValue, TagChild, css, div, tags
 
+from ...ui._tag import consolidate_attrs
 from ...ui._x._htmldeps import value_box_dependency
-from ...ui._x._utils import consolidate_attrs
 from ...ui.css_unit import CssUnit, as_css_unit, as_width_unit
 from ...ui.fill import as_fill_carrier
 from ._card import CardItem, card, card_body
